[BUGFIX] Make sure mime type is available when no property is specified
The TYPO3 indexer in
 sysext/core/Classes/Resource/Index/Indexer.php#gatherFileInformationArray
calls "getFileInfoByIdentifier" with an empty "$propertiesToExtract"
array, which means it wants to have all information.

S3 driver did not yet return the mime type in that case, causing a crash:

> PHP Warning: Undefined array key "mime_type"
> in typo3/sysext/core/Classes/Resource/Index/Indexer.php line 323

So instead of only fetching the MIME type when it is explicitly requested,
we also fetch it when no property has been given.

The error occured when a file was not in sys_file but already existed
in storage - e.g. when the storage had been filled outside of TYPO3.